requirements.txt
tdqm
matplotlib
openpyxl
yfinance
pandas
gradio
plyer
pyyaml
urllib